Fix issues sizing components  (#4452)
* Override box-sizing in jupyter

* Fix FileSelector and CrossSelector sizing

* Fix TextEditor sizing

* Fix Vega sizing

* Fix Tabulator theme handling

* Use random ids in notebooks

* Update Markup pane sizing

* Standardize widget sizing

* Fix TextEditor load warning

* Fix Perspective theming and streaming

* Minor doc updates

* Fix url prefixing on model update

* Apply resize callback on Vega

* Restore resize events on grid templates

* Small fix for ECharts sizing

* Remove FlexBox resize loop

* Improve Plotly sizing behavior

* Further ECharts sizing improvements

* Fix Select widget sizing

* Revert Image filetype change

* Fix unit tests

* Update unit tests

* Fix sizing of LoadingSpinner

* Fix ui tests

+++ panel/models/widgets.py
@@ -85,7 +85,7 @@ class Audio(HTMLBox):
 
     value = Any(help="Encoded file data")
 
-    volume = Nullable(Int, help="""The volume of the audio player.""")
+    volume = Nullable(Float, help="""The volume of the audio player.""")
 
 
 class Video(HTMLBox):
